Updated delegated harvesting guide
