+++ source/resources/examples/typescript/aggregate/CosigningAggregateBondedTransactionsAutomaticallyWithConstraints.js
@@ -26,7 +26,7 @@ const validTransaction = (transaction, publicAccount) => {
         transaction.signer.equals(publicAccount) &&
         transaction.mosaics.length === 1 &&
         (transaction.mosaics[0].id.equals(new nem2_sdk_1.MosaicId('75AF035421401EF0') ||
-            transaction.mosaics[0].id.equals(new nem2_sdk_1.NamespaceId('nem.xem')))) &&
+            transaction.mosaics[0].id.equals(new nem2_sdk_1.NamespaceId('symbol.xym')))) &&
         transaction.mosaics[0].amount.compare(nem2_sdk_1.UInt64.fromUint(100 * Math.pow(10, 6))) < 0;
 };
 const cosignAggregateBondedTransaction = (transaction, account) => {
