+++ source/resources/examples/typescript/aggregate/CreatingAnEscrowContractWithAggregateBondedTransaction.js
@@ -32,11 +32,11 @@ const ticketDistributorPublicAccount = nem2_sdk_1.PublicAccount.createFromPublic
 const ticketMosaicId = new nem2_sdk_1.MosaicId('7cdf3b117a3c40cc');
 // replace with ticket mosaic id divisibility
 const ticketDivisibility = 0;
-// replace with nem.xem id
+// replace with symbol.xym id
 const networkCurrencyMosaicId = new nem2_sdk_1.MosaicId('75AF035421401EF0');
 // replace with network currency divisibility
 const networkCurrencyDivisibility = 6;
-const aliceToTicketDistributorTx = nem2_sdk_1.TransferTransaction.create(nem2_sdk_1.Deadline.create(), ticketDistributorPublicAccount.address, [new nem2_sdk_1.Mosaic(networkCurrencyMosaicId, nem2_sdk_1.UInt64.fromUint(100 * Math.pow(10, networkCurrencyDivisibility)))], nem2_sdk_1.PlainMessage.create('send 100 nem.xem to distributor'), networkType);
+const aliceToTicketDistributorTx = nem2_sdk_1.TransferTransaction.create(nem2_sdk_1.Deadline.create(), ticketDistributorPublicAccount.address, [new nem2_sdk_1.Mosaic(networkCurrencyMosaicId, nem2_sdk_1.UInt64.fromUint(100 * Math.pow(10, networkCurrencyDivisibility)))], nem2_sdk_1.PlainMessage.create('send 100 symbol.xym to distributor'), networkType);
 const ticketDistributorToAliceTx = nem2_sdk_1.TransferTransaction.create(nem2_sdk_1.Deadline.create(), aliceAccount.address, [new nem2_sdk_1.Mosaic(ticketMosaicId, nem2_sdk_1.UInt64.fromUint(1 * Math.pow(10, ticketDivisibility)))], nem2_sdk_1.PlainMessage.create('send 1 museum ticket to customer'), networkType);
 /* end block 01 */
 /* start block 02 */
