+++ source/resources/examples/typescript/blockchain/GettingTheCurrentMosaicIdentifierBehindANamespace.js
@@ -21,7 +21,7 @@ const nem2_sdk_1 = require("nem2-sdk");
 const RepositoryFactoryHttp_1 = require("nem2-sdk/dist/src/infrastructure/RepositoryFactoryHttp");
 /* start block 01 */
 // replace with namespace name
-const namespace = new nem2_sdk_1.NamespaceId('nem.xem');
+const namespace = new nem2_sdk_1.NamespaceId('symbol.xym');
 // replace with node endpoint
 const nodeUrl = 'http://api-harvest-20.us-west-1.nemtech.network:3000';
 const repositoryFactory = new RepositoryFactoryHttp_1.RepositoryFactoryHttp(nodeUrl);
