+++ source/resources/examples/typescript/metadata/AssigningMetadataToAnAccount.js
@@ -44,7 +44,7 @@ const signedTransaction = bobAccount.sign(aggregateTransaction, networkGeneratio
 console.log(signedTransaction.hash);
 /* end block 03 */
 /* start block 04 */
-// replace with nem.xem id
+// replace with symbol.xym id
 const networkCurrencyMosaicId = new nem2_sdk_1.MosaicId('75AF035421401EF0');
 // replace with network currency divisibility
 const networkCurrencyDivisibility = 6;
