+++ source/resources/examples/typescript/multisig/ConvertingAnAccountToMultisig.js
@@ -45,7 +45,7 @@ const signedTransaction = account.sign(aggregateTransaction, networkGenerationHa
 console.log(signedTransaction.hash);
 /* end block 04 */
 /* start block 05 */
-// replace with nem.xem id
+// replace with symbol.xym id
 const networkCurrencyMosaicId = new nem2_sdk_1.MosaicId('75AF035421401EF0');
 // replace with network currency divisibility
 const networkCurrencyDivisibility = 6;
